Added support for using %if_dirs as operators
Last commit before unplugging the old nasm preprocessor.

Added support for using all the %if_dirs as arithmetic operators in PP
expressions. %ifdef and %ifndef, however, must be used with the {%ifdef
abc} syntax to avoid the expansion of %if, eg:
%if %ifdef abc
In the above the entire part after %if will be expanded before they are
processed as expression, so abc will be fully expanded to an undefined
object.

A %if_dir operator accepts only one argument. When multiple arguments
are needed, use the operator with a pair of parenthesis like this:
%if %ifidn(abc, def)